All query stuff is now done in a single parselet
