+++ Rant/Rant.csproj
@@ -204,13 +204,8 @@
     <Compile Include="Engine\Compiler\Parselets\ConstantLiteral.cs" />
     <Compile Include="Engine\Compiler\Parselets\DefaultParselet.cs" />
     <Compile Include="Engine\Compiler\Parselets\EscapeSequence.cs" />
-    <Compile Include="Engine\Compiler\Parselets\Hyphen.cs" />
     <Compile Include="Engine\Compiler\Parselets\LeftAngle.cs" />
-    <Compile Include="Engine\Compiler\Parselets\LeftParen.cs" />
     <Compile Include="Engine\Compiler\Parselets\Parselet.cs" />
-    <Compile Include="Engine\Compiler\Parselets\RegexFilter.cs" />
-    <Compile Include="Engine\Compiler\Parselets\RightAngle.cs" />
-    <Compile Include="Engine\Compiler\Parselets\Subtype.cs" />
     <Compile Include="Engine\Compiler\Parselets\Whitespace.cs" />
     <Compile Include="Engine\Compiler\RantExpressionCompiler.cs" />
     <Compile Include="Engine\Constructs\AttribPersistence.cs" />
